Cleanup and restructuring.
