+++ src/main.lua
@@ -23,7 +23,7 @@ end
 ---------------------------------------
 -- hump libraries here ----------------
 
-require("cupid.debug")
+require "cupid.debug"
 Timer = require "hump.timer"
 Vector = require "hump.vector"
 Camera = require "hump.camera"
@@ -53,6 +53,7 @@ local color_blue =	{30, 144, 255} -- dodger blue
 
 local chunk_height = 128
 
+	printf("[inc]: brain->mem[0]: %d\n", brain->mem[0]);
 function setup_game()
 	screen_w = love.graphics.getWidth()
 	screen_h = love.graphics.getHeight()
